Preliminary Shader
-Testing shader logic and VAO bindings

+++ src/ECS/Entity/Entity.h
@@ -3,11 +3,10 @@
 class Entity
 {
 private:
-	int ID;
+	unsigned long ID;
 
 public:
 	Entity(int ID) : ID(ID) {};
 	Entity(const Entity& entity) = default;
 	int GetID() const;
 };
-
